Add the "all-candidates" option.
Change the logic of whether output all candidates or the one with the closest m/z value. This is previously controlled by the "denovo" option. If "denovo" is on, all candidates will be saved. Otherwise, only the best. In this commit, an additional "all-candidates" option is added to control this behavior.